Move `getSubmitButtonLabel` to shared utils
